Major refactor: all settings now globally available via `setting()`
Rationale: "The road to programmer hell is paved with global
variables", but in some cases, "practicality beats purity". In this
instance, we already had several debates about whether it wouldn't
make more sense to make settings available globally instead of always
passing it around via function arguments.

Using function arguments means we can avoid a global variable and
limits access to the settings dict to those functions that really need
it. However, in practice, almost all function *do* need it. (And in a
few annoying cases, the parent function doesn't, but a function it
calls does.) Effectively, the settings dict was being passed to most
functions, which pretty much makes it a global variable anyway.

Therefore, I have now "taken it out of circulation" and stuck the
actual settings dict in a `let` block in `run_simulation.jl`. The way
to access it is now via `setting()`. This has simplified a lot of code.
An added advantage is that any changes to the settings dict by the software
have to be explicitly declared, as you now have to use
`initsettings()` or `updatesettings()` for that.